+++ order-service/src/main/java/com/logistic/order/adapter/in/web/response/CreateOrderResponse.java
@@ -1,6 +1,5 @@
-package com.logistic.order.adaptor.in.web.response;
+package com.logistic.order.adapter.in.web.response;
 
-import com.logistic.common.internal.response.OrderClientResponse;
 import java.util.List;
 
 public record CreateOrderResponse(
@@ -9,7 +8,7 @@ public record CreateOrderResponse(
     Long buyerId,
     String memo,
     String status,
-    List<OrderClientResponse.OrderProduct> orderProductList) {
+    List<OrderProduct> orderProductList) {
 
   public record OrderProduct(
       Long productId,
