+++ order-service/src/main/java/com/logistic/order/application/port/OrderPersistencePort.java
@@ -1,5 +1,6 @@
 package com.logistic.order.application.port;
 
+import com.logistic.order.application.port.in.query.SearchOrderQuery;
 import com.logistic.order.domain.Order;
 import java.util.Optional;
